+++ src/core/greeting.js
@@ -1,5 +1,6 @@
 import { version } from '../../package.json'
 import chalk from 'chalk'
+import { log } from './log'
 
 const ascii = `
                     __     __          __ 
@@ -9,11 +10,11 @@ const ascii = `
 /_/ /_/\\____/\\__,_/\\___/_.___/\\____/\\__/  `
 
 const info = `
-PACKAGE | Nodebot   v${version}
-AUTHOR  | trustgit  (c) ${new Date().getFullYear()} MIT License`
+ PACKAGE | Nodebot   v${version} 
+ AUTHOR  | trustgit  (c) ${new Date().getFullYear()} MIT License `
 
 export default function greeting() {
     const colors = ['red', 'green', 'yellow', 'blue', 'magenta', 'cyan', 'black']
-    console.log(chalk[colors[Math.floor(Math.random() * colors.length)]](ascii))
-    console.log(chalk.black.bgWhite(info))
+    log(chalk[colors[Math.floor(Math.random() * colors.length)]](ascii))
+    log(chalk.black.bgWhite(info))
 }